allow .atoum.yml file to be absent
In order to load the extension and still leave the user to choice to load
the .atoum.php or the .atoum.yml file, we need to avoid having an exception
when the file is not present.

So when catch the exception if the file has not been found in the folders.